system-install: instant reboot/shutdown for iso builds

+++ user/.config/scripts/poweroff.sh
@@ -1,6 +1,10 @@
 #!/bin/sh
 if command -v systemctl; then
-    systemctl poweroff
+    if [ -f /etc/iso ]; then
+        doas systemctl poweroff -ff
+    else
+        systemctl poweroff
+    fi
 elif command -v loginctl; then
     loginctl poweroff
 fi
