+++ user/.config/scripts/reboot.sh
@@ -1,7 +1,10 @@
 #!/bin/sh
 if command -v systemctl; then
-    echo a
-    systemctl reboot
+    if [ -f /etc/iso ]; then
+        doas systemctl reboot -ff
+    else
+        systemctl reboot
+    fi
 elif command -v loginctl; then
     loginctl reboot
 fi
